Fixed Position

+++ Position.java
@@ -5,6 +5,8 @@
  *
  */
 
+import java.math.BigDecimal;
+
 /**
  * Position is a class which defines a datatype Position which contains an x and y coord.
  *
@@ -15,15 +17,15 @@ public class Position {
 	 * the x coord for a position.
 	 * the y coord for a position.
 	 */
-	private double positionX;
-	private double positionY;
+	private BigDecimal positionX;
+	private BigDecimal positionY;
 	
 	/**
 	 * This defines a position of x and y.
 	 * @param x
 	 * @param y
 	 */
-	public Position (double x, double y) {
+	public Position (BigDecimal x, BigDecimal y) {
 		this.positionX = x;
 		this.positionY = y;
 	}
@@ -32,16 +34,17 @@ public class Position {
 	 * Returns the x and y coord as an array of doubleegers.
 	 *	@return 1D array consisting of x and y.
 	 */
-	public double[] getPosition () {
-		double retrunArray[] = {positionX, positionY};
+	public BigDecimal[] getPosition () {
+		BigDecimal retrunArray[] = {positionX, positionY};
 		return retrunArray;
 	}
 	
 	/**
 	 * Overwrites the current x and y of a position,
 	 */
-	public void setPosition (double x, double y) {
+	public void setPosition (BigDecimal x, BigDecimal y) {
 		this.positionX = x;
 		this.positionY = y;
 	}
 }
+
